*** empty log message ***

+++ dev-db/unixODBC/unixODBC-2.2.6.ebuild
@@ -1,6 +1,6 @@
-# Copyright 1999-2003 Gentoo Technologies, Inc.
+# Copyright 1999-2005 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-db/unixODBC/unixODBC-2.2.6.ebuild,v 1.1 2003/11/21 19:16:19 rphillips Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-db/unixODBC/unixODBC-2.2.6.ebuild,v 1.1.1.1 2005/11/30 10:11:40 chriswhite Exp $
 
 DESCRIPTION="ODBC Interface for Linux"
 HOMEPAGE="http://www.unixodbc.org/"
@@ -8,20 +8,21 @@ SRC_URI="http://www.unixodbc.org/${P}.tar.gz"
 
 LICENSE="GPL-2"
 SLOT="0"
-KEYWORDS="x86 ~ppc ~hppa ~alpha amd64"
-IUSE="qt"
+KEYWORDS="x86 ppc hppa alpha amd64 sparc ia64"
+IUSE="qt gnome"
 
-DEPEND="virtual/glibc
+DEPEND="virtual/libc
 	>=sys-libs/readline-4.1
 	>=sys-libs/ncurses-5.2
-	qt? ( >=x11-libs/qt-3.0* )"
+	qt? ( =x11-libs/qt-3* )
+	gnome? ( gnome-base/gnome-libs )"
 
 src_compile() {
 	local myconf
 
-	if [ "`use qt`" ]
+	if use qt
 	then
-		myconf="--enable-gui=yes"
+		myconf="--enable-gui=yes --x-libraries=/usr/lib --x-includes=/usr/include/X11"
 	else
 		myconf="--enable-gui=no"
 	fi
@@ -32,10 +33,38 @@ src_compile() {
 		    ${myconf} || die
 
 	make || die
+
+	if use gnome
+	then
+		# Symlink for configure
+		ln -s ${S}/odbcinst/.libs ./lib
+		# Symlink for libtool
+		ln -s ${S}/odbcinst/.libs ./lib/.libs
+		cd gODBCConfig
+		./configure --host=${CHOST} \
+				--with-odbc=${S} \
+				--prefix=/usr \
+				--x-libraries=/usr/lib \
+				--sysconfdir=/etc/unixODBC \
+				|| die
+
+		# not sure why these symlinks are needed. busted configure, i guess...
+		ln -s ../depcomp .
+		ln -s ../libtool .
+		make || die
+		cd ..
+	fi
 }
 
 src_install() {
 	make DESTDIR=${D} install || die
+
+	if use gnome
+	then
+		cd gODBCConfig
+		make DESTDIR=${D} install || die
+		cd ..
+	fi
 
 	dodoc AUTHORS COPYING ChangeLog NEWS README*
 	find doc/ -name "Makefile*" -exec rm '{}' \;
